+++ StabilityMatrix.Avalonia/ViewModels/SettingsViewModel.cs
@@ -36,7 +36,8 @@ public partial class SettingsViewModel : PageViewModelBase
             vmFactory.Get<MainSettingsViewModel>(),
             vmFactory.Get<InferenceSettingsViewModel>(),
             vmFactory.Get<AccountSettingsViewModel>(),
-            vmFactory.Get<UpdateSettingsViewModel>()
+            vmFactory.Get<UpdateSettingsViewModel>(),
+            vmFactory.Get<NotificationSettingsViewModel>()
         };
 
         CurrentPagePath.AddRange(SubPages);